pipeline/requirements.txt
Pillow>=10.0
numpy>=1.24
scikit-learn>=1.3
google-genai>=1.0
